Server/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
sentence-transformers==3.1.1
requests==2.32.3
google-generativeai==0.8.3
python-multipart==0.0.12